galleria/requirements.txt
#pip3 install -r requirements.txt
Pillow==7.2.0
django-widget-tweaks